feat(api file structure): Modified file structure of api, introduce routes folder  and reorganized code based on tables.
