+++ backend/api/services/UserService.ts
@@ -1,6 +1,6 @@
 // services/UserService.ts
-import { db } from '../db'; // reuse the shared instance
-import { users, organizationUser } from '../db/schema';
+import { db } from '../../db'; // reuse the shared instance
+import { users, organizationUser } from '../../db/schema';
 import { eq } from 'drizzle-orm';
 
 export async function getUserById(id: string) {
@@ -13,9 +13,11 @@ export async function getUsersByOrganization(organizationId: number) {
   const rows = await db.query.organizationUser.findMany({
     where: eq(organizationUser.organizationId, organizationId),
     with: {
-      user: true, 
+      usersInAuth: true,
     },
   });
 
-  return rows.map((row) => row.user);
+  return rows.map((row) => ({
+    users: row.usersInAuth, // Rename usersInAuth to users
+  }));
 }
